Refactor grid into a separate GridBuilder
